feat(projection-runtime): enforce runtime core/store boundaries (redemeine-tzx)
