+++ package.json
@@ -56,6 +56,7 @@
     "docs:install": "bun install --cwd website",
     "docs:dev": "bun run --cwd website start",
     "docs:build": "bun run --cwd website build",
+    "check:projection-runtime-boundaries": "node ./bin/check-projection-runtime-boundaries.js",
     "verify:workspace": "bunx turbo run typecheck build test --continue"
   },
   "workspaces": [
