+++ packages/projection-runtime-store-inmemory/src/index.ts
@@ -1,0 +1,2 @@
+export { InMemoryProjectionStore } from './InMemoryProjectionStore';
+export { InMemoryProjectionLinkStore } from './InMemoryProjectionLinkStore';
