+++ packages/projection-runtime-store-mongodb/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "@redemeine/projection-runtime-store-mongodb",
+  "version": "0.1.0",
+  "private": true,
+  "main": "src/index.ts",
+  "module": "src/index.ts",
+  "types": "src/index.ts",
+  "exports": {
+    ".": {
+      "types": "./src/index.ts",
+      "import": "./src/index.ts",
+      "require": "./src/index.ts",
+      "default": "./src/index.ts"
+    }
+  },
+  "scripts": {
+    "build": "tsc -p tsconfig.json",
+    "typecheck": "tsc -p tsconfig.json --noEmit",
+    "test": "bun test",
+    "lint": "bunx tsc -p tsconfig.json --noEmit"
+  },
+  "dependencies": {
+    "@redemeine/projection-runtime-core": "workspace:*"
+  }
+}
