+++ packages/projection/src/IProjectionLinkStore.ts
@@ -1,5 +1,1 @@
-export interface IProjectionLinkStore {
-  addLink(aggregateType: string, aggregateId: string, targetDocId: string): Promise<void> | void;
-  resolveTarget(aggregateType: string, aggregateId: string): Promise<string | null> | string | null;
-  removeLinksForTarget?(targetDocId: string): Promise<void> | void;
-}
+export type { IProjectionLinkStore } from '@redemeine/projection-runtime-core';
