+++ tsconfig.typedoc.json
@@ -8,6 +8,9 @@
       "@redemeine/kernel": ["packages/kernel/src/index.ts"],
       "@redemeine/mirage": ["packages/mirage/src/index.ts"],
       "@redemeine/projection": ["packages/projection/src/index.ts"],
+      "@redemeine/projection-runtime-core": ["packages/projection-runtime-core/src/index.ts"],
+      "@redemeine/projection-runtime-store-inmemory": ["packages/projection-runtime-store-inmemory/src/index.ts"],
+      "@redemeine/projection-runtime-store-mongodb": ["packages/projection-runtime-store-mongodb/src/index.ts"],
       "@redemeine/saga": ["packages/saga/src/index.ts"],
       "@redemeine/saga-runtime": ["packages/saga-runtime/src/index.ts"]
     }
@@ -18,6 +21,9 @@
     "packages/kernel/src/**/*.ts",
     "packages/mirage/src/**/*.ts",
     "packages/projection/src/**/*.ts",
+    "packages/projection-runtime-core/src/**/*.ts",
+    "packages/projection-runtime-store-inmemory/src/**/*.ts",
+    "packages/projection-runtime-store-mongodb/src/**/*.ts",
     "packages/saga/src/**/*.ts",
     "packages/saga-runtime/src/**/*.ts"
   ]
